data_simulator/requirements.txt
psycopg2-binary
pandas
numpy
python-dotenv
requests
